pyproject.toml
[tool.poetry]
name = "llamacloud-streamlit"
version = "0.1.0"
description = "Streamlit App Showcasing LlamaCloud's Capabilities"
authors = ["Sourabh Desai <sourabh@runllama.ai>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.12.8"
llama-index-embeddings-openai = "^0.3.1"
llama-index-indices-managed-llama-cloud = "^0.6.3"
streamlit = "^1.41.1"
llama-cloud = "^0.1.7"


[tool.poetry.group.dev.dependencies]
watchdog = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
